Intento2
Emmanuel Valencia Lopera

+++ homework/pregunta_09.py
@@ -26,5 +26,5 @@ def pregunta_09():
 
     """
     df = pd.read_csv("files/input/tbl0.tsv", sep="\t")
-    df["year"] = df["c3"].str[:4].astype(int)
+    df["year"] = df["c3"].str[:4]
     return df
